Changed: * Project config files have been updated with new versions of dependency libs * The project organization and project package have been changed from cc.kevinlee to kevinlee
