+++ .travis.yml
@@ -4,7 +4,7 @@ cache:
     - "$HOME/.gradle"
     - "$HOME/.m2/repository"
 jdk:
-  - oraclejdk8
+  - openjdk8
 script: mvn clean test
 script: gradle clean test jar
