+++ src/main/java/kevinlee/testosterone/ExpectedExceptionAssertions.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package cc.kevinlee.testosterone;
+package kevinlee.testosterone;
 
 import java.util.function.Function;
