+++ src/main/java/kevinlee/testosterone/Testosterone.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package cc.kevinlee.testosterone;
+package kevinlee.testosterone;
 
 import java.util.concurrent.Callable;
 import java.util.concurrent.atomic.AtomicInteger;
