+++ src/main/java/kevinlee/testosterone/ThrowableTestResultHandler.java
@@ -13,13 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package cc.kevinlee.testosterone;
+package kevinlee.testosterone;
+
 
 /**
  * @author Lee, SeongHyun (Kevin)
- * @version 0.0.1 (2014-08-26)
+ * @version 0.0.1 (2014-08-17)
  *
  */
-public interface Then<T> {
-  Then<T> then(final T t);
+public interface ThrowableTestResultHandler<T> {
+  <EX extends Throwable> void expect(ExpectedExceptionAssertions<EX> expectedExceptionAssertion);
+
+  Then<T> then(T then) throws Exception;
 }
